rename command and removed Explore Feed notification
